rejig tests
