+++ tests/bluemira/geometry/test_tools.py
@@ -385,7 +385,7 @@ class TestPointAlongWire:
         wire = make_polygon(coords)
         desired_len = np.sqrt(2.5)
 
-        point = find_point_along_wire_at_length(wire, desired_len)
+        point = wire_value_at(wire, distance=desired_len)
 
         np.testing.assert_almost_equal(point, [1.5, 0, 0.5], decimal=2)
 
@@ -393,7 +393,7 @@ class TestPointAlongWire:
         circle = make_circle(radius=1, axis=[1, 1, 1])
         semi_circle = make_circle(radius=1, axis=[1, 1, 1], end_angle=180)
 
-        point = find_point_along_wire_at_length(circle, np.pi)
+        point = wire_value_at(circle, distance=np.pi)
 
         np.testing.assert_almost_equal(point, semi_circle.end_point().T[0], decimal=2)
 
@@ -444,51 +444,3 @@ class TestConvexHullWires2d:
 
         with pytest.raises(ValueError):
             convex_hull_wires_2d([circle], 10, plane=bad_plane)
-
-
-class TestWireValueAt:
-    @classmethod
-    def setup_class(cls):
-        cls.square = make_polygon(
-            {"x": [0, 1, 1, 0], "y": 0, "z": [0, 0, 1, 1]}, closed=True
-        )
-        line = make_polygon([[0, 0, 1], [0, 0, 0], [0, 0, 0]])
-        semicircle = make_circle(
-            1, center=(1, 0, -1), start_angle=90, end_angle=270, axis=(0, 1, 0)
-        )
-        line2 = make_polygon([[1, 0, -2], [2, 0, -2]])
-        cls.mixed = BluemiraWire([line, semicircle, line2])
-
-    def test_square_alpha(self):
-        assert np.allclose(self.square.value_at(alpha=0.0), self.square.start_point)
-        assert np.allclose(self.square.value_at(alpha=0.25), np.array([1, 0, 0]))
-        assert np.allclose(self.square.value_at(alpha=0.5), np.array([1, 0, 1]))
-        assert np.allclose(self.square.value_at(alpha=0.75), np.array([1, 0, 0]))
-        assert np.allclose(self.square.value_at(alpha=1.0), self.square.end_point)
-
-    def test_square_distance(self):
-        length = self.square.length
-        assert np.allclose(self.square.value_at(distance=0.0), self.square.start_point)
-        assert np.allclose(
-            self.square.value_at(distance=0.25 * length), np.array([1, 0, 0])
-        )
-        assert np.allclose(
-            self.square.value_at(distance=0.5 * length), np.array([1, 0, 1])
-        )
-        assert np.allclose(
-            self.square.value_at(distance=0.75 * length), np.array([1, 0, 0])
-        )
-        assert np.allclose(self.square.value_at(distance=length), self.square.end_point)
-
-    def test_mixed_alpha(self):
-        assert np.allclose(self.mixed.value_at(alpha=0.0), np.array([0, 0, 0]))
-        assert np.allclose(self.mixed.value_at(alpha=0.5), np.array([0, 0, -1]))
-        assert np.allclose(self.mixed.value_at(alpha=1.0), np.array([2, 0, -2]))
-
-    def test_mixed_distance(self):
-        length = self.mixed.length
-        assert np.allclose(self.mixed.value_at(distance=0.0), np.array([0, 0, 0]))
-        assert np.allclose(
-            self.mixed.value_at(distance=0.5 * length), np.array([0, 0, -1])
-        )
-        assert np.allclose(self.mixed.value_at(distance=length), np.array([2, 0, -2]))
