+++ tests/bluemira/geometry/test_wire.py
@@ -22,6 +22,7 @@ import numpy as np
 
 from bluemira.geometry.coordinates import Coordinates
 from bluemira.geometry.tools import make_bezier, make_circle, make_polygon
+from bluemira.geometry.wire import BluemiraWire
 
 
 class TestWire:
@@ -80,3 +81,51 @@ class TestWire:
         start_point = wire.end_point()
 
         np.testing.assert_equal(start_point, np.array([[5], [0.2], [90]]))
+
+
+class TestWireValueAt:
+    @classmethod
+    def setup_class(cls):
+        cls.square = make_polygon(
+            {"x": [0, 1, 1, 0], "y": 0, "z": [0, 0, 1, 1]}, closed=True
+        )
+        line = make_polygon([[0, 0, 1], [0, 0, 0], [0, 0, 0]])
+        semicircle = make_circle(
+            1, center=(1, 0, -1), start_angle=90, end_angle=270, axis=(0, 1, 0)
+        )
+        line2 = make_polygon([[1, 0, -2], [2, 0, -2]])
+        cls.mixed = BluemiraWire([line, semicircle, line2])
+
+    def test_square_alpha(self):
+        assert np.allclose(self.square.value_at(alpha=0.0), self.square.start_point)
+        assert np.allclose(self.square.value_at(alpha=0.25), np.array([1, 0, 0]))
+        assert np.allclose(self.square.value_at(alpha=0.5), np.array([1, 0, 1]))
+        assert np.allclose(self.square.value_at(alpha=0.75), np.array([1, 0, 0]))
+        assert np.allclose(self.square.value_at(alpha=1.0), self.square.end_point)
+
+    def test_square_distance(self):
+        length = self.square.length
+        assert np.allclose(self.square.value_at(distance=0.0), self.square.start_point)
+        assert np.allclose(
+            self.square.value_at(distance=0.25 * length), np.array([1, 0, 0])
+        )
+        assert np.allclose(
+            self.square.value_at(distance=0.5 * length), np.array([1, 0, 1])
+        )
+        assert np.allclose(
+            self.square.value_at(distance=0.75 * length), np.array([1, 0, 0])
+        )
+        assert np.allclose(self.square.value_at(distance=length), self.square.end_point)
+
+    def test_mixed_alpha(self):
+        assert np.allclose(self.mixed.value_at(alpha=0.0), np.array([0, 0, 0]))
+        assert np.allclose(self.mixed.value_at(alpha=0.5), np.array([0, 0, -1]))
+        assert np.allclose(self.mixed.value_at(alpha=1.0), np.array([2, 0, -2]))
+
+    def test_mixed_distance(self):
+        length = self.mixed.length
+        assert np.allclose(self.mixed.value_at(distance=0.0), np.array([0, 0, 0]))
+        assert np.allclose(
+            self.mixed.value_at(distance=0.5 * length), np.array([0, 0, -1])
+        )
+        assert np.allclose(self.mixed.value_at(distance=length), np.array([2, 0, -2]))
